requirements.txt
python-dotenv>=1.0.0
openai>=1.30.0
openai-whisper>=20231117
picamera2>=0.3.19
openwakeword>=0.6.0
sounddevice>=0.4.6
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
pvporcupine>=4.0.0
